Refactor file functions from core/manager to core/utils
