+++ framework/wazuh/core/tests/data/utils/test_decoders.xml
