+++ framework/wazuh/core/tests/data/utils/test_rules.xml
