+++ framework/wazuh/core/tests/data/utils/test_rules_injection.xml
