+++ framework/wazuh/core/tests/test_manager.py
@@ -4,7 +4,7 @@
 # This program is a free software; you can redistribute it and/or modify it under the terms of GPLv2
 
 import os
-from unittest.mock import patch, mock_open
+from unittest.mock import patch
 
 import pytest
 
@@ -13,7 +13,6 @@ with patch('wazuh.core.common.ossec_uid'):
         from wazuh.core.manager import *
         from wazuh.core.exception import WazuhException
 
-ossec_cdb_list = "172.16.19.:\n172.16.19.:\n192.168.:"
 test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data', 'manager')
 ossec_log_path = '{0}/ossec_log.log'.format(test_data_path)
 
@@ -123,269 +122,6 @@ def test_get_logs_summary():
                    for log in result.values())
         assert result['wazuh-modulesd:database'] == {'all': 2, 'info': 0, 'error': 0, 'critical': 0, 'warning': 0,
                                                      'debug': 2}
-
-
-@patch('wazuh.core.manager.check_remote_commands')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_prettify_xml(mock_remote_commands, test_manager):
-    """Tests prettify_xml method works and methods inside are called with expected parameters"""
-    input_file, _ = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with open(os.path.join(test_data_path, input_file)) as f:
-        xml_file = f.read()
-    m = mock_open(read_data=xml_file)
-    with patch('builtins.open', m):
-        result = prettify_xml(xml_file)
-
-    assert isinstance(result, str)
-    mock_remote_commands.assert_called_once_with(result)
-
-
-@patch('wazuh.core.manager.check_remote_commands')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_prettify_xml_formula_injection(mock_remote_commands):
-    """Test that prettify_xml method avoids formula injections."""
-    with open(os.path.join(test_data_path, 'test_rules_injection.xml')) as f:
-        xml_file = f.read()
-    m = mock_open(read_data=xml_file)
-    with patch('builtins.open', m):
-        result = prettify_xml(xml_file)
-
-    assert isinstance(result, str)
-    description = re.search(r'<description>(.+)</description>', result)
-    assert description, 'Could not find description tag in XML file'
-    assert description.group(1).startswith("'"), f'Did not prepend formulas: {description}'
-
-
-@patch('time.time', return_value=0)
-@patch('random.randint', return_value=0)
-@patch('wazuh.core.manager.chmod')
-@patch('wazuh.core.manager.check_remote_commands')
-@patch('wazuh.core.manager.safe_move')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_upload_xml(mock_safe, mock_check_remote_commands, mock_chmod, mock_random, mock_time, test_manager):
-    """Tests upload_xml method works and methods inside are called with expected parameters"""
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with open(os.path.join(test_data_path, input_file)) as f:
-        xml_file = f.read()
-    m = mock_open(read_data=xml_file)
-    with patch('builtins.open', m):
-        result = upload_xml(xml_file, output_file)
-
-    assert isinstance(result, WazuhResult)
-    mock_time.assert_called_once_with()
-    mock_random.assert_called_once_with(0, 1000)
-    m.assert_any_call(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.xml'), 'w')
-    mock_chmod.assert_called_once_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.xml'), 0o660)
-    mock_safe.assert_called_once_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.xml'),
-                                      os.path.join(test_data_path, output_file),
-                                      permissions=0o660)
-
-
-@pytest.mark.parametrize('effect, expected_exception', [
-    (ExpatError, 1113)
-])
-def test_prettify_xml_open_ko(effect, expected_exception, test_manager):
-    """Tests prettify_xml function works when open function raise an exception
-
-    Parameters
-    ----------
-    effect : Exception
-        Exception to be triggered.
-    expected_exception
-        Expected code when triggering the exception.
-    """
-    input_file, _ = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with patch('wazuh.core.manager.load_wazuh_xml', side_effect=effect):
-        with pytest.raises(WazuhException, match=f'.* {expected_exception} .*'):
-            prettify_xml(input_file)
-
-
-@pytest.mark.parametrize('effect, expected_exception', [
-    (IOError, 1005)
-])
-def test_upload_xml_open_ko(effect, expected_exception, test_manager):
-    """Tests upload_xml function works when open function raise an exception
-
-    Parameters
-    ----------
-    effect : Exception
-        Exception to be triggered.
-    expected_exception
-        Expected code when triggering the exception.
-    """
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with patch('wazuh.core.manager.open', side_effect=effect):
-        with pytest.raises(WazuhException, match=f'.* {expected_exception} .*'):
-            upload_xml(input_file, output_file)
-
-
-@patch('time.time', return_value=0)
-@patch('random.randint', return_value=0)
-@patch('wazuh.core.manager.check_remote_commands')
-@patch('wazuh.core.manager.chmod')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_upload_xml_ko(mock_chmod, mock_remote_commands, mock_random, mock_time, test_manager):
-    """Tests upload_xml function exception works and methods inside are called with expected parameters"""
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with open(os.path.join(test_data_path, input_file)) as f:
-        xml_file = f.read()
-    m = mock_open(read_data=xml_file)
-    with patch('builtins.open', m):
-        with patch('wazuh.core.manager.load_wazuh_xml', side_effect=Exception):
-            with pytest.raises(WazuhException, match=f'.* 1113 .*'):
-                upload_xml(xml_file, output_file)
-
-        with patch('wazuh.core.manager.load_wazuh_xml'):
-            with patch('wazuh.core.manager.safe_move', side_effect=Error):
-                with pytest.raises(WazuhException, match=f'.* 1016 .*'):
-                    upload_xml(xml_file, output_file)
-
-    mock_time.assert_called_with()
-    mock_random.assert_called_with(0, 1000)
-    mock_chmod.assert_called_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.xml'), 0o660)
-
-
-@patch('wazuh.core.manager.validate_cdb_list', return_value=True)
-@patch('time.time', return_value=0)
-@patch('random.randint', return_value=0)
-@patch('wazuh.core.manager.chmod')
-@patch('wazuh.core.manager.safe_move')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_upload_list(mock_safe, mock_chmod, mock_random, mock_time, mock_validate_cdb, test_manager):
-    """Tests upload_list function works and methods inside are called with expected parameters"""
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    ossec_log_file = get_logs()
-
-    m = mock_open(read_data=ossec_log_file)
-    with patch('builtins.open', m):
-        result = upload_list(input_file, output_file)
-
-    assert isinstance(result, WazuhResult)
-
-    mock_validate_cdb.assert_called_once_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.txt'))
-    mock_time.assert_called_once_with()
-    mock_random.assert_called_once_with(0, 1000)
-    mock_chmod.assert_called_once_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.txt'), 0o640)
-    mock_safe.assert_called_once_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.txt'),
-                                      os.path.join(test_data_path, output_file),
-                                      permissions=0o660)
-
-
-@pytest.mark.parametrize('effect, expected_exception', [
-    (IOError, 1005)
-])
-def test_upload_list_open_ko(effect, expected_exception, test_manager):
-    """Tests upload_list function works when open function raise an exception
-
-    Parameters
-    ----------
-    effect : Exception
-        Exception to be triggered.
-    expected_exception
-        Expected code when triggering the exception.
-    """
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    with patch('wazuh.core.manager.open', side_effect=effect):
-        with pytest.raises(WazuhException, match=f'.* {expected_exception} .*'):
-            upload_list(input_file, output_file)
-
-
-@patch('time.time', return_value=0)
-@patch('random.randint', return_value=0)
-@patch('wazuh.core.manager.chmod')
-@patch('wazuh.core.manager.common.ossec_path', new=test_data_path)
-def test_upload_list_ko(mock_chmod, mock_random, mock_time, test_manager):
-    """Tests upload_list function exception works and methods inside are called with expected parameters"""
-    input_file, output_file = getattr(test_manager, 'input_rules_file'), getattr(test_manager, 'output_rules_file')
-
-    ossec_log_file = get_logs()
-
-    m = mock_open(read_data=ossec_log_file)
-    with patch('builtins.open', m):
-        with pytest.raises(WazuhException, match=f'.* 1800 .*'):
-            upload_list(ossec_log_file, output_file)
-
-        with patch('wazuh.core.manager.validate_cdb_list', return_value=False):
-            with pytest.raises(WazuhException, match=f'.* 1800 .*'):
-                upload_list(ossec_log_file, output_file)
-
-        with patch('wazuh.core.manager.validate_cdb_list', return_value=True):
-            with patch('wazuh.core.manager.safe_move', side_effect=Error):
-                with pytest.raises(WazuhException, match=f'.* 1016 .*'):
-                    upload_list(ossec_log_file, output_file)
-
-        mock_time.assert_called_with()
-        mock_random.assert_called_with(0, 1000)
-        mock_chmod.assert_called_with(os.path.join(test_manager.api_tmp_path, 'api_tmp_file_0_0.txt'), 0o640)
-
-
-@pytest.mark.parametrize('input_file', [
-    'input_rules_file',
-    'input_decoders_file',
-    'input_lists_file'
-])
-@patch('wazuh.core.common.ossec_path', test_data_path)
-def test_validate_xml(input_file, test_manager):
-    """Tests validate_xml function works
-
-    Parameters
-    ----------
-    input_file : str
-        Name of the input file.
-    """
-    input_file = getattr(test_manager, input_file)
-    with open(os.path.join(test_data_path, input_file)) as f:
-        xml_file = f.read()
-
-    with patch('builtins.open', mock_open(read_data=xml_file)):
-        result = validate_xml(input_file)
-        assert result is True
-
-
-def test_validate_xml_ko():
-    """Tests validate_xml function exceptions works"""
-    # Open function raise IOError
-    with patch('wazuh.core.manager.open', side_effect=IOError):
-        with pytest.raises(WazuhException, match=f'.* 1005 .*'):
-            validate_xml('test_path')
-
-    # Open function raise ExpatError
-    with patch('wazuh.core.manager.open', side_effect=ExpatError):
-        result = validate_xml('test_path')
-        assert result is False
-
-
-def test_validate_cdb_list():
-    """Tests validate_cdb function works"""
-    m = mock_open(read_data=ossec_cdb_list)
-    with patch('builtins.open', m):
-        assert validate_cdb_list('path')
-
-
-@patch('wazuh.core.manager.re.match', return_value=False)
-def test_validate_cdb_list_ko(mock_match):
-    """Tests validate_cdb function exceptions works"""
-    # Match error
-
-    ossec_log_file = get_logs()
-
-    m = mock_open(read_data=ossec_log_file)
-    with patch('wazuh.core.manager.open', m):
-        result = validate_cdb_list('path')
-
-    assert result is False
-
-    # Open function raise IOError
-    with patch('wazuh.core.manager.open', side_effect=IOError):
-        with pytest.raises(WazuhException, match=f'.* 1005 .*'):
-            validate_cdb_list('path')
 
 
 @pytest.mark.parametrize('error_flag, error_msg', [
