+++ framework/wazuh/tests/test_decoder.py
@@ -187,7 +187,7 @@ def test_get_decoder_file_exceptions():
 @patch('wazuh.decoder.copyfile')
 @patch('wazuh.decoder.remove')
 @patch('wazuh.decoder.safe_move')
-@patch('wazuh.core.manager.check_remote_commands')
+@patch('wazuh.core.utils.check_remote_commands')
 def test_upload_file(mock_remote_commands, mock_safe_move, mock_remove, mock_copyfile, mock_xml, mock_delete, file, overwrite):
     """Test uploading a decoder file.
 
@@ -216,7 +216,7 @@ def test_upload_file(mock_remote_commands, mock_safe_move, mock_remove, mock_cop
 @patch('wazuh.decoder.delete_decoder_file')
 @patch('wazuh.decoder.upload_xml')
 @patch('wazuh.decoder.safe_move')
-@patch('wazuh.core.manager.check_remote_commands')
+@patch('wazuh.core.utils.check_remote_commands')
 def test_upload_file_ko(mock_remote_commands, mock_safe_move, mock_xml, mock_delete):
     """Test exceptions on upload function."""
     # Error when file exists and overwrite is not True
